agent/requirements.txt
# ============================================================================
# Sentinel API Requirements
# ============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP Client
httpx==0.25.1

# Database
asyncpg==0.29.0

# Environment
python-dotenv==1.0.0

# Cryptography (for artifact signing)
cryptography==41.0.7

# Optional: Development dependencies
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# mypy==1.7.1
